requirements.txt
# Writon - AI Text Processor Requirements
# Core dependencies (existing CLI)
requests==2.31.0
python-dotenv==1.0.0

# API dependencies (new FastAPI)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Development and testing (optional)
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1